+++ cli/src/lib/errors/errors.go
@@ -1,4 +1,4 @@
-package lib
+package errors
 
 func CheckError(err error) {
 	if err != nil {
