+++ cli/src/lib/fs/fs.go
@@ -1,6 +1,7 @@
-package lib
+package fs
 
 import (
+	"horahora/cli/src/lib/errors"
 	"io/fs"
 	"os"
 )
@@ -8,7 +9,7 @@ import (
 // Reads the folder at provided location.
 func ReadFolder(dirPath string) []fs.DirEntry {
 	folderContent, err := os.ReadDir(dirPath)
-	CheckError(err)
+	errors.CheckError(err)
 
 	return folderContent
 }
@@ -17,7 +18,7 @@ func ReadFolder(dirPath string) []fs.DirEntry {
 // and returns its content as a string.
 func ReadFile(filePath string) string {
 	data, err := os.ReadFile(filePath)
-	CheckError(err)
+	errors.CheckError(err)
 
 	return string(data)
 }
@@ -26,5 +27,5 @@ func ReadFile(filePath string) string {
 func WriteFile(filePath string, content string) {
 	fileContent := []byte(content)
 	err := os.WriteFile(filePath, fileContent, 0644)
-	CheckError(err)
+	errors.CheckError(err)
 }
