+++ cli/src/lib/json/json.go
@@ -1,13 +1,16 @@
-package lib
+package json
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"horahora/cli/src/lib/errors"
+)
 
 // Parses json string and returns the value of provided type.
 func ParseJSON[outputType any](inputJSON string) outputType {
 	jsonContent := []byte(inputJSON)
 	var jsonResult outputType
 	err := json.Unmarshal(jsonContent, &jsonResult)
-	CheckError(err)
+	errors.CheckError(err)
 
 	return jsonResult
 }
@@ -15,7 +18,7 @@ func ParseJSON[outputType any](inputJSON string) outputType {
 // Turns a provided value into a json string.
 func StringifyJSON[inputType any](inputItem inputType) string {
 	jsonString, err := json.Marshal(inputItem)
-	CheckError(err)
+	errors.CheckError(err)
 
 	return string(jsonString)
 }
@@ -23,7 +26,7 @@ func StringifyJSON[inputType any](inputItem inputType) string {
 // Turns a provided value into a human-readable json string.
 func PrettyJSON[inputType any](inputItem inputType) string {
 	prettyJSONString, err := json.MarshalIndent(inputItem, "", "  ")
-	CheckError(err)
+	errors.CheckError(err)
 
 	return string(prettyJSONString)
 }
